LayoutTests/platform/ios-simulator/css3/flexbox/flex-rounding-expected.txt

FAIL:
Expected 50 for width, but got 51. 

<div class="flexbox" style="width: 101px; height: 10px">
    <div data-expected-width="51" data-offset-x="0"></div>
    <div data-expected-width="50" data-offset-x="51"></div>
</div>
FAIL:
Expected 34 for width, but got 33. 

<div class="flexbox" style="width: 100px; height: 10px">
    <div data-expected-width="33" data-offset-x="0"></div>
    <div data-expected-width="34" data-offset-x="33"></div>
    <div data-expected-width="33" data-offset-x="67"></div>
</div>
FAIL:
Expected 50 for width, but got 51. 

<div class="flexbox" style="-webkit-flex-direction: row-reverse; width: 101px; height: 10px">
    <div data-expected-width="50" data-offset-x="51"></div>
    <div data-expected-width="51" data-offset-x="0"></div>
</div>
FAIL:
Expected 34 for width, but got 33. 

<div class="flexbox" style="-webkit-flex-direction: row-reverse; width: 100px; height: 10px">
    <div data-expected-width="33" data-offset-x="67"></div>
    <div data-expected-width="34" data-offset-x="33"></div>
    <div data-expected-width="33" data-offset-x="0"></div>
</div>
FAIL:
Expected 50 for height, but got 51. 

<div class="flexbox" style="-webkit-flex-direction: column; height: 101px; width: 100px">
    <div data-expected-height="51" data-offset-y="0"></div>
    <div data-expected-height="50" data-offset-y="51"></div>
</div>
FAIL:
Expected 34 for height, but got 33. 

<div class="flexbox" style="-webkit-flex-direction: column; height: 100px; width: 100px">
    <div data-expected-height="33" data-offset-y="0"></div>
    <div data-expected-height="34" data-offset-y="33"></div>
    <div data-expected-height="33" data-offset-y="67"></div>
</div>
FAIL:
Expected 50 for height, but got 51. 

<div class="flexbox" style="-webkit-flex-direction: column-reverse; height: 101px; width: 100px">
    <div data-expected-height="50" data-offset-y="51"></div>
    <div data-expected-height="51" data-offset-y="0"></div>
</div>
FAIL:
Expected 34 for height, but got 33. 

<div class="flexbox" style="-webkit-flex-direction: column-reverse; height: 100px; width: 100px">
    <div data-expected-height="33" data-offset-y="67"></div>
    <div data-expected-height="34" data-offset-y="33"></div>
    <div data-expected-height="33" data-offset-y="0"></div>
</div>
